requirements.txt
# jupyter notebook
jupyter
# pandas
pandas
# matplotlib
matplotlib
# pylint
pylint
# test
pytest
pytest-cov
nbval
# formatter
black
# rust based linter
ruff
# pyspark
pyspark
# micro service
fastapi
uvicorn[standard]
# Databricks
databricks-cli
databricks-connect==14.2.0
# Azure
azure-storage-blob
pyodbc